requirements.txt
accelerate==0.25.0
bitsandbytes @ https://github.com/jllllll/bitsandbytes-windows-webui/releases/download/wheels/bitsandbytes-0.41.1-py3-none-win_amd64.whl#sha256=adacda9b2b65dcb1931b222dffd7a91f0b611b3410d5b51c37ef7b22654106e6
colorama
pydantic==1.10.13
python-dotenv==1.0.0
PyYAML==6.0.1
transformers==4.36.2
torch==2.1.2+cu118
